Rename: stat 모듈 폴더 구조 변경
