+++ src/api/stat/stat.controller.ts
@@ -1,5 +1,5 @@
 import { Controller } from '@nestjs/common';
-import { StatService } from '../service/stat.service';
+import { StatService } from './stat.service';
 
 @Controller('stat')
 export class StatController {
